Version 8
This version will ask you 10 questions, telling you if you are correct between each question (if you are incorrect it will tell you what the correct answer was). Then it will bring you to an activity which will allow you to see your score, then return you to the main screen. The questions are imported from the triviaQuestions.txt file and are different from each other.

Changes: 
- Added feedback to the results page (if you get the question wrong it tells you what the correct answer was)

The files provided:
-From the src folder of the project and include the original code (mainActivity.java, Question.java, and Results.java), as well as some generated by android studio.
-The compiled application which can be run on an emulator (for best results please use a Pixel virtual machine running API30) app-debug.apk
-The zip folder containing all the code needed to load the project in android studio